# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "Code With Prince",
      "version": "1.0.0",
      "title": "Code With Prince Flask API Tutorial",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "tags": [
      {
        "name": "Products API",
        "description": "Test API"
      }
    ],
     "servers": [
          {
          "url": "/"
          }
      ],
         "components": {
          "schemas": {
              "store": {
                  "type": "object",
                  "properties": {
                      "title": {
                      "type": "string"
                      },
                      "herf": {
                      "type": "string"
                      },
                      "image": {
                        "type": "string"
                        },
                      "rating": {
                            "type": "string"
                            },
                        "price": {
                            "type": "string"
                            },
                        "location": {
                            "type": "string"
                            },
                        "amenities": {
                            "type": "string"
                            }
                  }
              }
                
          }
      },
       "paths": {
          
          "/search":{
            "post": {
                "tags": [
                    "Add todo to API"
                ],
                "summary": "Add todo",
                "requestBody": {
                "description": "Post New todo To Database",
                "required": true,
                "content": {
                    "application/json": {
                    "schema": {
                        "$ref": "/"
                        }
                    }
                }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                        "$ref": "#/components/schemas/store"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                        "$ref": "#/components/schemas/store"
                        }
                }
                }
                },
  
              "get" :{
                  "tags": [
                      "Get todos from API"
                  ],
                  "summary": "Get todos",
                  "parameters": [{
                    "name": "title",
                    "in": "query",
                    "description": "Give hotel's name",
                    "required": false,
                    "type": "string",
                    "schema" : {
                        "$ref": "#/components/schemas/store/properties/title"
                    }
                  },
                  {
                    "name": "location",
                    "in": "query",
                    "description": "Give hotel's location",
                    "required": false,
                    "type": "string",
                    "schema" : {
                        "$ref": "#/components/schemas/store/properties/location"
                    }
                  },
                  {
                    "name": "amenities",
                    "in": "query",
                    "description": "Give hotel's amenities",
                    "required": false,
                    "type": "string",
                    "schema" : {
                        "$ref": "#/components/schemas/store/properties/amenities"
                    }
                  },
                  {
                    "name": "price",
                    "in": "query",
                    "description": "Give price",
                    "required": false,
                    "type": "string",
                    "schema" : {
                        "$ref": "#/components/schemas/store/properties/price"
                    }
                  },
                  {
                    "name": "sorting",
                    "in": "query",
                    "description": "price(asc for Ascending or dsc for Descending)",
                    "required": false,
                    "type": "string",
                    "schema" : {
                        "$ref": "#/components/schemas/store/properties/price"
                    }
                  }
                ],
                 
                  "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                      "$ref": "#/components/schemas/store"
                      }
                  },
                  "404": {
                      "description": "NOT FOUND",
                      "schema": {
                      "$ref": "#/components/schemas/store"
                      }
              }
              }
              },
             
              "put": {
              "tags": [
                  "Update todos on API"
              ],
              "parameters": [{
                  "name": "title",
                  "in": "path",
                  "description": "todo id to update",
                  "required": true,
                  "type": "integer",
                  "schema" : {
                      "$ref": "#/components/schemas/store"
                  }
              }],
              "summary": "Update Todo",
              "requestBody": {
              "description": "Update Existing todo On Database",
              "required": true,
              "content": {
                  "application/json": {
                      "schema": {
                          "$ref": "#/components/schemas/store"
                      }
                  }
              }
              },
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                      "$ref": "#/components/schemas/store"
                      }
                  },
                  "404": {
                      "description": "NOT FOUND",
                      "schema": {
                      "$ref": "#/components/schemas/store"
                      }
              }
              }
          },
          "delete": {
              "tags": [
                  "Delete todo to API"
              ],
              "parameters": [{
                  "name": "title",
                  "in": "path",
                  "description": "product id to delete",
                  "required": true,
                  "type": "integer",
                  "schema" : {
                      "$ref": "#/components/schemas/store"
                  }
              }],
              "summary": "Delete todo",
              
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                      "$ref": "#/components/schemas/store"
                      }
                  },
                  "404": {
                      "description": "NOT FOUND",
                      "schema": {
                      "$ref": "#/components/schemas/store"
                      }
              }
              }
          }
          }
  }
  }